+++ src/level/RLModsNotesPopup.hpp
@@ -10,11 +10,13 @@ public:
 
 private:
     bool init() override;
+    void onNoteLabelClick(CCObject* sender);
 
     GJGameLevel* m_level = nullptr;
     cue::ListNode* m_listNode = nullptr;
     ScrollLayer* m_scrollLayer = nullptr;
     async::TaskHolder<web::WebResponse> m_getNotesTask;
+    std::vector<std::string> m_notes;
 
     ~RLModsNotesPopup() {
         m_getNotesTask.cancel();